Stay in device menu after changing settings
Functionality:
Change behaviour of device_menu to return to the last edited device's menu.
Add a Back-Selection to the device menu to go back to the main menu.
Add an Exit-Selection to the device menu to exit from there as well.

Code:
Made the divider a constant as it's now used in more than one place.