updated README
Added bc as a dependency and included a basic configuration example for using the script with Waybar and Hyprland. I think that adding the latter is important because  rofi-bluetooth can also work with rofi-wayland (a Wayland-compatible fork of Rofi), but there was no configuration provided for a Wayland environment. I thought it might be helpful to include one for users running Wayland setups.